% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdf.R
\name{npdf_cox}
\alias{npdf_cox}
\title{Estimating a Cox model with a shared nonparametric frailty}
\usage{
npdf_cox(data, time, status, family, covariates, population = 2,
  eps_conv = 10^-4)
}
\arguments{
\item{data}{a data.frame in which there are status, time, family and covariates variables}

\item{time}{name of the variable which indicates the status}

\item{status}{name of the variable which indicates the status}

\item{family}{name of the variable by which the data are divided}

\item{covariates}{vector of covariates' names}

\item{population}{initial number of latent populations}

\item{eps_conv}{tolerance}
}
\value{
This fuction returns a list of elements, whose length is equal to the initial number of populations + 4. The last four elements of the list show the best model according to loglikelihood, BIC and AIC, while the fourth term represents the minimum number of latent population for which each population has at least one member.
}
\description{
This function estimates the parameter of a semiparametric Cox model with a nonparametric frailty term.
}
\examples{
N <- 100
S <- 40
Lambda_0_inv = function( t, c=0.01, d=4.6 ) ( t^( 1/d ) )/c
beta <- 1.6
p <- c( 0.8, 0.2 )
w_values <- c( 0.8, 1.6 )
data <- simulWeibDiscreteFrailCovNPbaseInv( N, S, beta, Lambda_0_inv, p, w_values)

test_res <- npdf_cox( data, time = 'time', status = 'status', family = 'family', covariates = c('x'), population = 4, eps_conv=10^-4)
best_model_llik <- test_res[[5]] #population+1
best_model_BIC <- test_res[[6]] #population+2
best_model_AIC <- test_res[[7]] #population+3
best_model_K <- test_res[[8]] #population+4

test_res[[best_model_BIC]]
}
