% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.npdf.R
\name{plot.npdf}
\alias{plot.npdf}
\title{Survival curves from a fitted nonparametric discrete frailty model}
\usage{
\method{plot}{npdf}(x, survfit_opts = NULL, ...)
}
\arguments{
\item{x}{A fitted nonparametric discrete frailty model, as returned by \code{\link{npdf_cox}} with \code{estK=FALSE}}

\item{survfit_opts}{List of arguments to pass to \code{\link{survfit.formula}}}

\item{...}{Arguments to pass to \code{\link{plot.survfit}}}
}
\description{
This function plots Kaplan-Meier estimates of the survival for each fitted latent population.
}
\examples{
## TODO

}
