% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.npdf.R
\name{plot.npdflist}
\alias{plot.npdflist}
\title{Survival curves from a nonparametric discrete frailty model chosen by automatic model selection}
\usage{
\method{plot}{npdflist}(x, K = NULL, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{npdf_cox}} with \code{estK=TRUE}, containing a list of fitted nonparametric discrete frailty models}

\item{K}{The number of latent populations which identifies the specific model to plot survival curves for.  By default this is the model selected by the criterion specified when calling \code{\link{npdf_cox}}.}

\item{K}{The number of latent populations which identifies the specific model to plot survival curves for.  By default this is the model selected by the criterion specified when calling \code{\link{npdf_cox}}.}
}
\description{
Survival curves from a nonparametric discrete frailty model chosen by automatic model selection
}
\examples{
## TODO

}
