% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npdf.R
\name{print.npdf}
\alias{print.npdf}
\title{Print output from a fitted nonparametric discrete frailty model}
\usage{
\method{print}{npdf}(x, digits = NULL, ...)
}
\arguments{
\item{x}{A fitted nonparametric discrete frailty model, as returned by \code{\link{npdf_cox}} with \code{estK=FALSE}}

\item{digits}{Number of significant digits to present, by default \code{max(1, getOption("digits") - 4)}}

\item{...}{Further options (currently unused)}
}
\description{
Prints estimates, standard errors, likelihood and model comparison statistics from a fitted nonparametric discrete frailty model
}
