% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npdf.R
\name{print.npdflist}
\alias{print.npdflist}
\title{Print output from a nonparametric discrete frailty modelling procedure with automatic model selection.}
\usage{
\method{print}{npdflist}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{npdf_cox}} with \code{estK=TRUE}, containing a list of fitted nonparametric discrete frailty models}

\item{digits}{Number of significant digits to present, by default \code{max(1, getOption("digits") - 4)}}

\item{...}{Further options (currently unused)}
}
\description{
Print output from a nonparametric discrete frailty modelling procedure with automatic model selection.
}
