% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_weib.R
\name{simulWeibDiscreteFrail}
\alias{simulWeibDiscreteFrail}
\title{Generating time-to-event data with a Weibull baseline}
\usage{
simulWeibDiscreteFrail(N, S = NULL, lambda, rho, beta, p, w_values)
}
\arguments{
\item{N}{number of groups/clusters in data}

\item{S}{number of statistical units in each group}

\item{lambda}{value of lambda in a Weibull distribution}

\item{rho}{value of rho in a Weibull distribution}

\item{beta}{vector of regression parameters}

\item{p}{vector of proportions related to latent populations}

\item{w_values}{vector of frailty values}
}
\value{
This fuction returns a dataset which fits a parametric Cox model with a Weibull distributed baseline and a multiplicative discrete frailty term
}
\description{
Generating time-to-event data with a Weibull baseline
}
\examples{
N <- 100
S <- 40
lambda <- 0.5
beta <- 1.6
rho <- 1
p <- c( 0.8, 0.2 )
w_values <- c( 0.8, 1.6 )
data <- simulWeibDiscreteFrail( N, S, lambda, rho, beta, p, w_values)
head( data )

}
