% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_np.R
\name{simulWeibDiscreteFrailCovNPbaseInv}
\alias{simulWeibDiscreteFrailCovNPbaseInv}
\title{Generating time-to-event data with a nonparametric baseline}
\usage{
simulWeibDiscreteFrailCovNPbaseInv(N, S = NULL, beta, Lambda_0_inv, p,
  w_values)
}
\arguments{
\item{N}{number of groups/clusters in data}

\item{S}{number of statistical units in each group}

\item{beta}{vector of regression parameters}

\item{Lambda_0_inv}{inverse cumulative baseline function}

\item{p}{vector of proportions related to latent populations}

\item{w_values}{vector of frailty values}
}
\value{
This fuction returns a dataset which fits a semiparametric Cox model with a multiplicative discrete frailty term.
}
\description{
Generating time-to-event data with a nonparametric baseline
}
\examples{
N <- 100
S <- 40
Lambda_0_inv = function( t, c=0.01, d=4.6 ) ( t^( 1/d ) )/c
beta <- 1.6
p <- c( 0.8, 0.2 )
w_values <- c( 0.8, 1.6 )
data <- simulWeibDiscreteFrailCovNPbaseInv( N, S, Lambda_0_inv, rho, beta, p, w_values)
head( data )
}
