% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_np.R
\name{simulWeibDiscreteFrailCovNPbaseInv}
\alias{simulWeibDiscreteFrailCovNPbaseInv}
\title{Simulation of grouped time-to-event data with nonparametric baseline hazard and discrete shared frailty distribution}
\usage{
simulWeibDiscreteFrailCovNPbaseInv(N, S = NULL, beta, Lambda_0_inv, p,
  w_values)
}
\arguments{
\item{N}{number of groups in the data}

\item{S}{number of individuals in each group}

\item{beta}{vector of log hazard ratios}

\item{Lambda_0_inv}{inverse cumulative baseline hazard function, that is, with covariate values 0 and frailty ratio 1}

\item{p}{vector of K elements.  The kth element gives the proportion of groups in the kth latent population of groups.}

\item{w_values}{vector of K distinct frailty values, one for each latent population.}
}
\value{
A data frame with one row for each simulated individual, and the following columns: 

\code{family}: the group which the individual is in (integers 1, 2, ...)

\code{time}: the simulated event time 

\code{status}: the simulated survival status.   Censoring times are generated from a normal distribution with mean given by the simulated event time, and variance given by the mean divided by 10.  The event time is observed if it is less than the censoring time, and censored otherwise.

\code{x}: matrix of covariate values, generated from a standard normal distribution.

\code{belong}:  the frailty hazard ratio corresponding to the the cluster of groups in which the individual's group has been allocated.
}
\description{
This function returns a dataset generated from a semiparametric proportional hazards model with a shared discrete frailty term, for given cumulative baseline hazard function, hazard ratios, distribution of groups among latent populations, frailty values for each latent population, and randomly-generated covariate values.
}
\examples{
N <- 100
S <- 40
Lambda_0_inv = function( t, c=0.01, d=4.6 ) ( t^( 1/d ) )/c
beta <- 1.6
p <- c( 0.8, 0.2 )
w_values <- c( 0.8, 1.6 )
data <- simulWeibDiscreteFrailCovNPbaseInv( N, S, beta, Lambda_0_inv, p, w_values)
head( data )
}
